+++ src/main/java/com/dekk/global/security/oauth2/handler/OAuth2SuccessHandler.java
@@ -26,7 +26,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 
     private final String defaultRedirectUri;
     private final String joinPageUri;
-
     private final int accessTokenMaxAge;
     private final int refreshTokenMaxAge;
 
@@ -38,7 +37,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
             @Value("${app.oauth2.join-page}") String joinPageUri,
             @Value("${jwt.access-token-validity-in-seconds}") int accessTokenMaxAge,
             @Value("${jwt.refresh-token-validity-in-seconds}") int refreshTokenMaxAge) {
-
         this.jwtTokenProvider = jwtTokenProvider;
         this.refreshTokenRepository = refreshTokenRepository;
         this.cookieUtil = cookieUtil;
@@ -52,7 +50,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
     public void onAuthenticationSuccess(
             HttpServletRequest request, HttpServletResponse response, Authentication authentication)
             throws IOException, ServletException {
-
         String accessToken = jwtTokenProvider.createAccessToken(authentication);
         String refreshToken = jwtTokenProvider.createRefreshToken(authentication);
 
@@ -62,7 +59,6 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         String targetUrl = defaultRedirectUri;
 
         if (authentication.getPrincipal() instanceof CustomUserDetails userDetails) {
-
             refreshTokenRepository.save(RefreshToken.create(userDetails.getId(), refreshToken));
 
             if (userDetails.getStatus() == UserStatus.PENDING) {
